feat: modify transition logic and seperate logic (#218)
